rename files to have consistent naming
