+++ src/utils/ViewUtils.tsx
@@ -36,9 +36,9 @@ export function Logo(position: string = "relative") {
 export function Score(score: number, maxScore: number) {
   return (
     <div className="relative md:absolute top-0 right-0">
-        <h1 className={"p-4 text-2xl md:ml-5 text-center md:text-right" + blackText}>
-          Score: {score} / {maxScore}
-        </h1>
-      </div>
+      <h1 className={"p-4 text-2xl md:ml-5 text-center md:text-right" + blackText}>
+        Score: {score} / {maxScore}
+      </h1>
+    </div>
   )
 }
