+++ src/views/PlaylistSelectionView.tsx
@@ -1,5 +1,5 @@
 
-import { blackText } from "./ViewUtils";
+import { blackText } from "../utils/ViewUtils";
 
 export function PlaylistSelectionView(props: any) {
   return (
